+++ front/lib/actions/types/index.ts
@@ -58,3 +58,11 @@ export abstract class BaseAction {
     model: ModelConfigurationType;
   }): Promise<FunctionMessageTypeModel>;
 }
+
+export type ExtractActionBlob<T extends BaseAction> = Pick<
+  T,
+  {
+    // eslint-disable-next-line @typescript-eslint/ban-types
+    [K in keyof T]: T[K] extends Function ? never : K;
+  }[keyof T]
+>;
