+++ front/lib/models/assistant/actions/browse.ts
@@ -1,7 +1,7 @@
 import type { CreationOptional, ForeignKey } from "sequelize";
 import { DataTypes } from "sequelize";
 
-import type { BrowseActionOutputType } from "@app/lib/actions/types/browse";
+import type { BrowseActionOutputType } from "@app/lib/actions/browse";
 import { AgentConfiguration } from "@app/lib/models/assistant/agent";
 import { AgentMessage } from "@app/lib/models/assistant/conversation";
 import { frontSequelize } from "@app/lib/resources/storage";
